refactor: add sklearn model as base class
