+++ pi_optimal/models/sklearn/linear_model.py
@@ -1,8 +1,8 @@
-from .base_model import BaseModel
+from .base_sklearn_model import BaseSklearnModel
 from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
 from sklearn.linear_model import LinearRegression, LogisticRegression
 
-class LinearModel(BaseModel):
+class LinearModel(BaseSklearnModel):
     def __init__(
         self,
     ):
